pyproject.toml
[project]
name = "discord-bot-rewrite"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "discord-py>=2.7.1",
    "dotenv>=0.9.9",
    "openwebui-client>=0.3.2",
    "requests>=2.33.0",
]
